+++ src/globals/index.ts
@@ -277,6 +277,7 @@ export type IdevsExportOptions = IdevsExportRequest & {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   grid?: DataGrid<any, any>
   service: string
+  dialogTitle?: string
 }
 
 export type ExportOptions = IdevsExportOptions & {